i2c-gpio: Fixed PORTD initialization bug
PIC executs read and modify write when programmed with "PORTxbits.Rx =
0".  However, if PORT is in input mode, it will overwrite the current
state of the port.

When writing the program below, if RD0 is in input mode and the input
level is High, RD0 will be set to "1".

PORTDbits.RD0 = 0;
PORTDbits.RD1 = 0;

RD0 and RD1 are set to 0 at system startup. So I removed the
initialization with get_i2c ().