Makefile: Remove wildcard from CF
Having $(wildcard) make the list dynamic, meaning that when not found
it won't be in the list.  That's file as long as a file is the same
name as the previous run.  But it's not, if a different name is used.

Remve wildcard function from CF so that the patterns will always
present at clean target.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>