usart: Move object allocation to usart.c
tx_msg and rx_msg are used by both main.c and usart.c.  We allocate
them in usart.c and let main.c know we have them and accessible.

newline[] is only used by send_msg().  Thus moved to it.  This make it
allocated on the stack. So make sure we have enough stack.

rx_msg_delimiter is only used in receive_msg_int().  This variable is
changed to define so that now memory location for the variable is
needed.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>