i2c: Remove FPGA knowledge from I2C
Currently FPGA knowlege is spread all over the code base in TRCH. The
I2C driver is a low level communication driver, which should be
implemented without knowing how the bus and pins on the board are
connected. The idea of the board is a bit higher than low level i2c
driver.

This commit pulls out about FPGA and pin configurations from the i2c
driver.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>