ina3221: Convert type of the parameter type
The type parameter for ina3221_data_read() describes either shunt
voltage or bus voltage of INA3221 register.  Make the type of it to
enumeration so that we can explicitly specify shunt or bus voltage,
instead of 0 or 1.

While at it, this commit refactors the register address calculation to
to_reg().

REG_VOLTAGE_BASE doesn't have to be exported to the other
modules. Thus, it's moved to ina3221.c.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>