timer: Rename gtimer and hide it from outside
gtimer is a 32 bit variable updated by the timer interrupt handler for
every second.  On 8 bit PIC, accessing a 32 bit value on memory is not
atomic, thus need to disable the interrupt.

We offer a convenient function timer_get_gtimer, or now
timer_get_ticks, for users of the 32 bit values.

The variable is now static in timer.c and renamed to current_ticks. No
access to the variable is allowed except through the accessor.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>