ina3221: Make ina3221_data_read return -1 when anything goes wrong
The previous behavior of ina3221_data_read() is a bit strange.

 - It returns 0 when it can access I2C bus, 1 otherwise
 - When I2C NAKed, it returns 0 but set the error member variable to 1

With this commit, the function returns -1 (yes negative value) when
anything goes wrong, i2c bus access, programming error, or anything.

The error member variable is set to a reason why it failed. The
possible values are defined in the header file.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>