trch.h: Rename FPGA_CDRST_B to FPGA_FPGA_PWR_EN
RA5, named FPGA_CDRST_B, was defined but not used.  This pin is now
connected to a DC/DC (IC31)'s EN pin (2nd pin) on A001.  Rename it to
FPGA_PWR_EN, which is how it's called on A1001.

While working on the file, remove the trailing white space.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>