trch.h: Define FPGA_PWR_EN and FPGA_PWR_EN_DIR
FPGA_PWR_EN is connected from the Port A bit 5 (RA5) of TRCH, which is
pin 24, to the 1.0 V DC/DC on A1001.  The DC/DC is the start of the
FPGA power sequence.

The RA5 is bi-directional port, which means that we must control the
direction register, TRISA5, as well.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>