fpga: Implement new state machine
With EM1, the power of the FPGA was always ON.  Thus, we must control
the configuration of the FPGA by INIT_B and PROGRAM_B.

ince A1001, however, the power of the FPGA is controlled by the TRCH
with FPGA_PWR_EN for lower power consumption.  This allows us not to
use INIT_B and PROGRAM_B in the state machine.  Both pins are set to
INPUT in trch_init().

This commit implements a new simpler state machine.

- power off: Wait for config_ok = 1
- ready: Wait for VDD_3V3 = 1
- config: Wait for FPGA_WATCHDOG = 1
- active: Monitor FPGA_WATCHDOG toggle

In all states, when user ask to shutdown the FPGA by setting config_ok
to 0, it power off the FPGA and go back to power_off state.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>