main: Add fpga_program_maybe
Having watchdog is a good thing but it makes development process
tedious. It's because once the FPGA stops kicking the WDT, TRCH will
power cycle the FPGA, which might be during a device programming.

Moreover, the OBC has two flash memory chips for the FPGA
configuration data. The chip selection is done by TRCH and TRCH
selects the first chip after PoR. So there is nothing to do when you
are programming the first chip.

If, however, you want to program the second chip, you must make sure
the second chips is selected. In order to do so, you tell TRCH using a
FPGA signal, FPGA_CFG_MEM_SEL. This signal is dedicated for the FPGA
to request TRCH which chip to activate for the next PoR.

For device programming, TRCH will monitor the signal FPGA_CFG_MEM_SEL
and set TRCH_CFG_MEM_SEL to 1 when the signal become one. It does NOT
and must not set back to 0, because the signal will become HiZ while
programming.  TRCH must keep TRCH_CFG_MEM_SEL either 0 or 1 as
requested by the FPGA.

This functionality effectively disables all nominal operation on TRCH.
We highly recommend removing it from the final flight software.  You
can set -DCONFIG_FPGA_PROGRAM_MODE=0 in the Makefile to remove it.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>