trch.h: Define port dir and port level
On PIC16F87X, port direction is OUT when you write 0 to TRIS register.
I used to work with SoC with direction registers, which go OUT when
you write 1.  It's just too confusing to remember this.  So, I decided
to define it.

I didn't really need to define LOW and HIGH as 0 and 1 but just for
the sake of completeness.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>