fpga: Fix FPGA_INIT_B
FPGA_INIT_B is PORT A bit 4, which is Open-Drain.  Setting the port
OUT HIGH is same as setting IN.  In either setting the port become HiZ
and the signal will be pull-up by the external pull-up resistor.

Only in trans_to_config(), we've used OUT HIGH to mean HiZ.  In all
other transition functions, we are using IN to mean HiZ.  Unify the
code and use the same IN to mean HiZ.

No functionality has been changed.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>