Add IS_ENABLED() macro
We use several defines to control compile time features.  But our
usage is not clear whether we should define it empty or 1.  If an
empty define is used in if condition, it errors out.

IS_ENABLE() allows us to use the defines either with if-condition C
syntax and / or the pre-processor #ifdef.

Unfortunately this macro generates a warning like:

   src/fpga.c:34:: warning: (759) expression generates no code

This can be suppressed by using

   #pragma warning disable 759

But it's a warning generated by the code generator and the pragma
disables the all warning in the compilation unit. Thus we are not
disabling it for now.

To quote 50002737F MPLAB XC8 C Compiler User’s Guide for PIC® MCU:
> Specific instances of a warning produced by the code generator cannot
> be individually controlled and the pragma will remain in force during
> compilation of the entire C program, even across modules.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>