timer: Introduce MSEC_TO_TICKS
We can calculate a millisecond to ticks.  In order to wait for a
certain period, we are currently specifying time in ticks.  It's
been OK because we've been using 1 Hz tick.  But it seems we need
finer jiffies for system management.

This macro allows us to specify the time in micro second instead of
system dependent ticks.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>