fpga: Convert the type of activate_fpga to enum
This commit is preparation of the next commit.  We want to add another
value to activate_fpga.  To do so we have to replace Boolean to enum.

This commit converts Boolean type to a new enum type and maps false to
0 and true to 1, which are declared as FPGA_SHUTDOWN and FPGA_ACTIVATE
respectively, so that all the logic works without any fix.

No functionality should have been changed.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>