Remove ifdef on fpga_management_data
Having #ifdef to remove members from a struct causes build error
because the compiler assume that the member is still in the struct due
to IS_ENABLE() macro.

Remove #ifdef for a workaround.  This adds a few unused bytes to
constraint environment.  We'll fix this issue after the v1.0
release.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>