main: Remove sensor acquisitions
Having a large struct is not a best way for small device like PIC.
And we are not using these acquired values right now.

The code in the main is just wrapping around actual i2c. Remove these
code from main for now and we will re-introduce it later on.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>