trch: Replace literal port values
It's nice that we have documented all port directions and levels. But
it's quite tough to keep the documentation and code aligned.  In fact,
we already have a value mismatch for SPICAN_CS_B.  The doc says
SPICAN_CS_B is OUT HIGH, but the actual value is OUT LOW.

Instead, make code readable as documentation. Define OUT/IN and
LOW/HIGH for TRIS and PORT values and convert all definitions with
them.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>