ioboard: Add IO boards build support
Each project should have its own IO boards.  src/ioboard is the
directory for them.  If the directory has a ioboard.mk, included it.
The make file should define IOBOARD_SRCS with its source files.
The files will be appended to SRCS and built.

Because the source files are not in src/ but under src/ioboard/, it
can't access the header files in src/.  Add "-I src" to the compile
command.

Because the current XC8 with C90 library doesn't support weak
attribute, this commit adds a empty ioboard_init() as a library.  The
library, libioboard.a, must be linked very last.

Signed-off-by: Yasushi SHOJI <yashi@spacecubics.com>