zephyr/drivers/can: Implement TX API
This commit implement can_send API for Space Cubics CAN.

In the current FPA CAN IP core, CAN Packets stored in the TX Buffer
transmit to the CAN Bus in a FIFO (First In First Out).
Therefore, there is no issue with managing the list of TX callbacks
using an array.
However, due to future modifications in the FPGA CAN IP core, the
transmission order of CAN Packets stored in the TX buffer will be
rearranged based on the CAN ID (priority).
So, we will need to change the management of TX callbacks from an
array to something like an slist.

And SC CAN Controller notify the interrupt every times while ACK
error detected, so this driver received the many interrupt.
we also plan to improve the Interrupt Status Register

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>