boards: Add QSPI definition for MAIN board
This commit adds the QSPI definition for MAIN board.

SC-SAT1 MAIN board has four QSPI control cores, each of which is
connected to the following devices:

Configuration Flash is equipped with two chips, but due to the
commonization of SPI signal lines, it appears as a single memory
to the driver.

  - spi0
    1) Configuration Flash 1 or 2 (Selected by System Register)

  - spi1
    1) Data Store Flash 1
    2) Data Store Flash 2

  - spi2
    1) FRAM 1
    2) FRAM 2

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>