main,adcs: Enable SPI
This commit adds the `CONFIG_SPI` for MAIN and ADCS boards

And, we have added the necessary CONFIG_SPI_EXTENDED_MODES for
future support of QSPI and other features, but as of now,
it is not yet implemented in the driver.

Additionally, adding the SPI driver in the out-of-tree driver
has led to the following warnings during the build process

  No SOURCES given to Zephyr library: drivers__spi

This reasons is that the device driver directories under the
zephyr proper are activated in order to build the board specific
device drivers but no driver source code under the zephyr proper is
selected.  There is a workaround: The sensor device driver, for
example, has

  zephyr_library_property(ALLOW_EMPTY TRUE)

in the CMakeLists.txt.  Unfortunately, CMakeLists.txt under spi
don't have it.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>