zephyr: Add SC I2C driver
This commit adds Space Cubics I2C driver, but only implement the
`i2c_sc_transfer` API.

We have not yet implemented the `i2c_sc_configure` API. Therefore,
the communication speed of I2C is fixed at the default value of the
I2C IP core, which is Standard-mode(100 Kbps).

And I2C IP core supports a 10-bit address, but it has not been
implemented in this driver yet.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>