zephyr/can: Fix TX errors
During the continuous testing of CSP ping/uptime, the following
error occurred.

  TX callback is not registerd

This occurs when the TX done interrupt happens too early before
incrementing the tx_head in the TX callback managed by the
CAN driver.

When writing to TX Register4 in the FPGA IP core is completed,
it is necessary to increment the tx_head in the TX callback
first before transmitting the packet onto the CAN bus.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>