bootloader: Add simple bootloader
This commit adds a simple bootloader.
This simple bootloader directly accesses the registers of the FPGA
SPI Core without using the SPI driver in Zephyr.
These code is based on the code from the PDI repository.

All settings such as the address of the config memory where the
Zephyr application is stored and the size to be copied need to be
configured in Kconfig.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>