hwtest/adcs: Add get syshk command to CSP
We are currently facing an issue where the depletion of the CPS
buffer occurs when forwarding system HK from the ADCS board to
the ground using the router fuction of the MAIN board.

Since the cause of this issue is still unknown, we will implement
a workaround to return the latest ADCS System HK in response to
requests triggered from the MAIN board, without utilizing the
router function.

In this commit, we implement the functionality to store and return
the latest System HK. Note that System HK is stored in a fixed-size
FIFO, but currently only the latest one is returned.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>