hwtest/tools: Do not connect plot during blank time
In the current Yamcs telemetry parser, even during long periods of
absence of data, the plot lines are connected, making it difficult
to identify data gaps.
With this commit, we will not connect the lines between plots when
there is a gap in the data for 60 seconds.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>

add eps param