hwtest/main: Add DSTRX-3 downlink loop test
This commit adds the loop test for DSTRX-3 downlink.
This command writes 1 byte of data to the TX buffer of DSTRX-3 and
downlinks it every second. There are two arguments: the first is the
number of loops for the downlink, and the second is a control flag
for the downlink operation. Need to specific the bitwise OR of the
following values for the control flag.

  - b100000: NRZI Enable
  - b010000: Convolutional Code Enable
  - b001000: Pseudo-randomized Enable
  - b000100: Reed Solomon Enable
  - b000010: CRC32C Enable
  - b000001: Zero Padding Enable

for example, If you want to enable Read solmon and Zero padding,
please specify 5 (b000101).

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>