boards: Add NOR flash dts to MAIN/ADCS boards
This commit adds the DTS for the NOR flash (S25FL256L) to the MAIN/
ADCS boards. In the configuration data area, it adds the definitions
 for the `bit partition`, which includes the FPGA/bootloader, and
the `fsw partition`, which includes the Zephyr FSW, although the
sizes are provisional.

This commit only adds the NOR flash for configuration data because
the current SPI NOR flash driver supports only a single instance.

And, the following issue is helpful in understanding why the
`data-capture-mode` is necessary when using the SPI NOR flash driver.

  - https://github.com/zephyrproject-rtos/zephyr/issues/45441
  - https://github.com/zephyrproject-rtos/zephyr/issues/27843

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>