dts,bootloader: Change config memory partition
From SC-Sat1 FPGA hash e640ddcc, a mechanism will be implemented to
prepare golden and update images in a single config memory area,
allowing for fallback. Therefore, the config area will first be
partitioned into four sections.

In this commit, the bootloader is fixed to load the FSW for the
update image, but this will be corrected in a subsequent commit.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>