Revert "drivers/misc: Remove zephyr_library()"
This reverts commit 9658d19d4e4a86228569bb9c5e7100f93271bf27.

The following commit message described in 9658d19d is incorrect.

  `The misc driver for SC-Sat1 does not need to enable Zephyr's
   built-in misc driver. Therefore, we will remove zephyr_library().`

The zephyr_library() function is used to declare a library based on
the current directory. If we add source code without calling
zephyr_library(), it will be included in libzephyr.a.
While this does not impact the Zephyr build itself, it is not
ideal to include out-of-tree drivers in libzephyr.a, which is why
we are reverting this commit.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>