driver/misc: Kill waring when build the misc driver
Currently, when we build the ADCS board, the following warning
message appears:

  CMake Warning at ...
    No SOURCES given to Zephyr library: ..__scsat1-fsw__drivers__misc

This is because zephyr_library() is being called unconditionally in
misc/CMakefile, so for ADCS boards that do not use DSTRX-3, the
drivers under the misc directory are not included in the build.

Therefore, similar to the misc drivers under the Zephyr directory,
we will create a subdirectory under misc and modify the configuration
so that the subdirectory is added only when the relevant driver is
enabled in the Kconfig.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>