drivers/misc/dstrx3: Remove meaningless init process
In the current driver initialization process, HK acquisition from
DSTRX-3 and parameter setting to DSTRX-3 are performed. However,
since the DSTRX-3 is not powered on at the time of driver
initialization, this initialization process is meaningless.

After powering on the DSTRX-3, the appropriate settings should be
performed from the application side.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>