tests/hwtest/main: Change destination port for ADCS telemetry
In the following commit, we are implementing a workaround to send
ADCS telemetry to the ground. However, since we are reusing the CSP
connection to retrieve telemetry from the ADCS, the destination port
changes each time. In this commit, we modify it to explicitly send
to the same destination port as MAIN telemetry.

  - 285246
    hwtest/main: Change the downlink method for ADCS

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>