bootloader: Change target board
In the following commit, the target board for the bootloader is
changed to MAIN board, but the MAIN board's DTS contains settings
that are unnecessary for the ADCS board's bootloader. Therefore,
to define only the common settings for both MAIN and ADCS boards,
the target board will be changed to `scobc_module1`, and only the
necessary definitions will be added to app.overlay.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>