drivers/spi: Fix RX/TX buffer overflow on LittleFS
When introducing LittleFS to the SPI NOR flash, 64 bytes of data
were being transmitted and received at once to the SPI device.
However, since the QSPI Core's TX/RX buffer is only 16 bytes, it was
causing an overflow. This commit limits the data transmission and
reception to the SPI device to 16 bytes at a time.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>