main,adcs: Enable LittleFS for data storage
This commit enables the file system for data storage using LittleFS.
Currently, only a single partition named `/storage` is created, but
it may be split into multiple partitions in the future. And the Man
stack size is increased to 4096 bytes. This is aligned with the
configuration values from the LittleFS sample.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>

stack size