lib/csp: Add Upload (DATA) command
This commit adds the write data command needed for file upload from
the Ground System to SC-Sat1. This command includes the session ID
indicating the file opened with OPEN command, the write offset, the
size, and the actual datae.

This command is expected to be sent continuously, so sending a reply
telemetry each time would waste downlink bandwidth and could cause
collisions with the upload command on the CAN bus. Therefore, a
mechanism has been implemented to stack a certain number of results
and return a response only when a set threshold is reached.

The maximum length of the data to be sent can be specified in Kconfig.
Currently, it is set to 200 bytes, considering the uplink frame limit
of the SRS-3.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>