lib/flash: Add eraseing API for Config NOR flash
The commit adds an erasing API for Config NOR flash.
Since erasing the Config NOR flash is not a frequently performed
operation, it will adopt a specification that uniformly disallows
concurrent execution, regardless of differences in banks or
partitions. And If both offset and size are 0, the entire area will
be erased.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>