lib/csp: Add erase command handler for Config NOR flash
This commit adds a command handler to erase partitions on Config NOR
flash. Since operations on NOR flash are often time-consuming, the
actual operation is executed on a work queue to prevent blocking the
CSP handler's read processing. Although a work queue is used, only a
single queue is allocated for NOR flash operations; if another
operation is already in progress, the command is rejected. This
approach is intended to simplify the implementation.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>