main,adcs: Enhancement of power command
If an abnormal packet size or command code is received, the reply
telemetry was not sent previously. This commit addresses this by
ensuring the reply telemetry is sent. In such cases, the command
ID will be uniformly set to 0xFF, as this is necessary for
categorization on the Yamcs side.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>