drivers/misc/fpga: Add ECC Error Collect API
This commit adds an API to enable or disable ECC Error Collect. When
this feature is enabled, the HRMEM IP core performs a write-back of
the corrected data to repair the memory upon detection of a 1-bit ECC
error. This feature is enabled by default in the HRMEM IP core.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>