drivers/misc/fpga: Add Memory Scrub API
This commit adds an API to enable or disable memory scrub. When this
feature is enabled, the HRMEM IP core performs periodic reads of all
fields in the SRAM to check the written data.This feature is enabled
by default in the HRMEM IP core.

The HRMEM IP core also allows configuration of settings such as the
read cycle and behavior when the FIFO for error correction data
becomes full. However, this API only controls the enable/disable
functionality, keeping the default settings unchanged.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>