main,adcs: Add System HK sending worker skeleton
This commit adds a skeleton for a worker to send System HK to the
ground. The worker uses `k_timer` and `k_work` to run periodically,
with the sending interval configurable in Kconfig. The work queue
has a dedicated queue, and the stack size and priority can also be
configured in Kconfig. Additionally, there is a required no-sending
period after satellite deployment, and this period is also
configurable via Kconfig. At this stage, the content of System HK has
not yet been generated.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>