main,adcs: Add Kconfig for automatic downlink.
This commit adds a setting in Kconfig to control whether System HK
is automatically downlinked. The default setting is `y` for MAIN and
`n` for ADCS. This is because the ADCS will not be activated during
initial operations, and telemetry will be obtained in a
`request-reply` way even after activation.

This setting is planned to be changeable via command in the future,
but for now, it can only be configured in Kconfig.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>