main,adcs: Add Temperature to syshk
This commit adds a worker to monitor temperature periodically. The
monitor uses `k_timer` and `k_work` to periodically retrieve
temperature data from the sensor and publish it to subscribers via
Zbus. Subscribers monitor the temperature data channel and copy the
data into the System HK structure. The System HK structure is
prepared with a buffer solely for storing temperature sensor data.

In the future, we may implement a table defining offsets and sizes
for copying, similar to the cFS HK application, to make the structure
independent of its members.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>