dts: Remove SPI definition for FRAM
This commit removes the unused SPI definitions for FRAM in DTS. In
subsequent commits, we will add a library for FRAM, but it will be
temporarily removed from SPI driver management as it directly
operates the registers of the FPGA QSPI core. In the future, an
IP core dedicated to FRAM access is planned to be implemented.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>