lib/fram: Add FRAM read/write API
This commit adds APIs for reading from and writing to FRAM. However,
these APIs support only up to 16 bytes for each read and write
operation at a time. Additionally, the implementation of this library
does not use the SPI driver but instead directly operates the
registers of the FPGA QSPI core. This code is based on a simplified
version previously written for testing purposes.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>