main: Enable ADCS power
This commit enables the power switch for the ADCS Board during the
startup of the MAIN Board. However, since the power supply from the
EPS, which is the power source, is set to OFF by default, the ADCS
Board will not start automatically. This commit is intended to
activate the ADCS Board by simply turning ON the power channel from
the EPS to the ADCS.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>