lib/csp: Add copy command to Config NOR flash
This commit adds a command to copy files uploaded to Data Storage to
the Configuration NOR Flash in order to update the FPGA and FSW on
SC-Sat1. Since the FPGA image is large, up to 10MB, the file is
expected to be uploaded in segments, with an offset adjustment to
write to the NOR Flash. Additionally, to allow partial copying from
the middle of a file, the file offset can also be specified.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>