lib/flash: Workaround of Config NOR bank1 issue
SC-Sat1 has two Configuration NOR flash memory banks (bank0/bank1).
Normally, FPGA configuration is performed from bank0, and the FSW
(Flight Software) also boots from the same bank.

However, when an FSW booted from bank0 executes operations such as
erasing NOR flash on bank1, the device appears to ignore the erase
command.

This issue occurs because the NOR flash used in SC-Sat1 has a
capacity of 32MB, requiring the 4-byte address access mode to be
configured on the NOR flash. However, this mode is not set for the
NOR flash on the bank1 side.

The 4-byte address mode is set only during the initialization of the
spi-nor driver. When the FSW is booted from bank0, the NOR flash on
the disconnected bank1 side does not have this setting applied.

As a workaround for this issue, this commit introduces initialization
processing in the FSW that configures the 4-byte address mode for
both bank0 and bank1.

As a permanent solution, a broader review, including updates to the
spi-nor driver, should be considered to determine the best approach
to resolve this issue.

 Issue: #99

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>