lib/csp: Increase parameters for File Upload feature
Currently, during File Upload, we have faced that when data packets
are missing, and LittleFS attempts to write data to an offset larger
than the current file size, a process is triggered to copy previously
written block data to a new block. This copying process can take up
to about 7 seconds.

During this time, CSP packets accumulate in the libcsp buffer without
being released, causing the application to stuck. We will investigate
the root cause of the stucking issue further, but as a workaround, we
will increase the FIFO-related settings in libcsp and th
eapplication's work queue size.

These values are provisional for testing purposes. However, even with
this commit applied, memory usage remains at around 11%.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>