lib/fram: Add update and read API for file CRC
Since CRC32 calculation may take time, it might not be completed
within an AOS period. To address this, an API will be added to store
the CRC32 calculation result in the FRAM of MAIN/ADCS boards. The FRAM
storage is limited to a single entry, allowing only the result of the
last CRC32 calculation to be saved.

Signed-off-by: Takuya Sasaki <takuya.sasaki@spacecubics.com>